+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Improved fetch instances endpoint, now it also fetch other instances even if they are not connected
 * Added conversion of audios for sending recorded audio, now it is possible to send mp3 audios and not just ogg
 * Route to fetch all groups that the connection is part of
+* Route to fetch all privacy settings
 
 ### Fixed